+++ lib/index.js
@@ -1,4 +1,9 @@
-var lodash = require("lodash");
+var lodash = {
+  assign: require("lodash.assign"),
+  isObject: require("lodash.isobject"),
+  clone: require("lodash.clone"),
+  keys: require("lodash.keys")
+};
 var defaultValidators = require("./validators");
 var defaultTransforms = require("./transforms");
 var validate = require("./validate");
