+++ lib/transforms.js
@@ -1,4 +1,6 @@
-var lodash = require("lodash");
+var lodash = {
+  isNumber: require("lodash.isnumber")
+};
 
 function notNullOrUndefined (value) {
   return value !== null && value !== undefined;
@@ -47,76 +49,6 @@ module.exports = {
       return true;
     }  else if (value === false || /^(false|off|no)$/.test(value)) {
       return false;
-    } else {
-      return value;
-    }
-  },
-  capitalize: function (value) {
-    if (lodash.isString(value)) {
-      return lodash.capitalize(value);
-    } else {
-      return value;
-    }
-  },
-  compact: function (value) {
-    if (lodash.isArray(value)) {
-      return lodash.compact(value);
-    } else {
-      return value;
-    }
-  },
-  deburr: function (value) {
-    if (lodash.isString(value)) {
-      return lodash.deburr(value);
-    } else {
-      return value;
-    }
-  },
-  escape: function (value) {
-    if (lodash.isString(value)) {
-      return lodash.escape(value);
-    } else {
-      return value;
-    }
-  },
-  kebabCase: function (value) {
-    if (lodash.isString(value)) {
-      return lodash.kebabCase(value);
-    } else {
-      return value;
-    }
-  },
-  toLowerCase: function (value) {
-    if (lodash.isString(value)) {
-      return value.toLowerCase();
-    } else {
-      return value;
-    }
-  },
-  trim: function (value) {
-    if (lodash.isString(value)) {
-      return lodash.trim(value);
-    } else {
-      return value;
-    }
-  },
-  toUpperCase: function (value) {
-    if (lodash.isString(value)) {
-      return value.toUpperCase();
-    } else {
-      return value;
-    }
-  },
-  uniq: function (value) {
-    if (lodash.isArray(value)) {
-      return lodash.uniq(value);
-    } else {
-      return value;
-    }
-  },
-  words: function (value) {
-    if (lodash.isString(value)) {
-      return lodash.words(value);
     } else {
       return value;
     }
