+++ lib/validate.js
@@ -1,4 +1,16 @@
-var lodash = require("lodash");
+var lodash = {
+  isObject: require("lodash.isobject"),
+  keys: require("lodash.keys"),
+  forIn: require("lodash.forin"),
+  contains: require("lodash.contains"),
+  isNumber: require("lodash.isnumber"),
+  isString: require("lodash.isstring"),
+  isDate: require("lodash.isdate"),
+  isBoolean: require("lodash.isboolean"),
+  isNull: require("lodash.isnull"),
+  isArray: require("lodash.isarray")
+}
+
 
 /*
  * get the type of the value
@@ -55,7 +67,7 @@ module.exports = function validate (testSchema, value, output, error, validatorO
 
   // Recurse into any arrays and build the validated output
   if (lodash.isArray(value) && testSchema.schema) {
-    var arrayResults = lodash.map(value, function(arrayValue) {
+    var arrayResults = value.map(function(arrayValue) {
       return validate(
         testSchema.schema,
         arrayValue,
@@ -64,10 +76,10 @@ module.exports = function validate (testSchema, value, output, error, validatorO
         validatorObj
       );
     });
-    output = lodash.map(arrayResults, function (rslt) {
+    output = arrayResults.map(function (rslt) {
       return rslt.data;
     });
-    error = lodash.map(arrayResults, function (rslt) {
+    error = arrayResults.map(function (rslt) {
       return rslt.error;
     });
 
@@ -116,7 +128,7 @@ module.exports = function validate (testSchema, value, output, error, validatorO
     // Check the type
     var type = getType(output);
     var allowedTypes = [].concat(testSchema.type);
-    if (testSchema.type && !lodash.include(allowedTypes, type)) {
+    if (testSchema.type && !lodash.contains(allowedTypes, type)) {
       isValid = false;
       error.type = "'" + type + "' was not in allowed types: " + allowedTypes.join(", ");
     }
